修改抽水验证矿机还是否掉线

+++ src/util/mod.rs
@@ -192,8 +192,8 @@ pub async fn get_pool_stream_with_tls(
 }
 
 // 根据抽水率计算启动多少个线程
-pub fn clac_phread_num(rate: f32) -> i32 {
-    (rate * 1000.0) as i32
+pub fn clac_phread_num(rate: f32) -> u64 {
+    (rate * 1000.0) as u64
 }
 
 #[test]
